+++ core/src/rest/futopt/intraday/ticker.rs
@@ -62,10 +62,7 @@ impl<'a> TickerRequestBuilder<'a> {
         }
 
         // Make request
-        let request = self.client.agent().get(&url);
-        let request = self.client.auth().apply_to_request(request);
-
-        let response = self.client.execute(request)?;
+        let response = self.client.get(&url)?;
         let ticker: FutOptTicker = crate::rest::read_json(response)?;
 
         Ok(ticker)