+++ core/src/rest/stock/corporate_actions/capital_changes.rs
@@ -73,10 +73,7 @@ impl<'a> CapitalChangesRequestBuilder<'a> {
         }
 
         // Make request
-        let request = self.client.agent().get(&url);
-        let request = self.client.auth().apply_to_request(request);
-
-        let response = self.client.execute(request)?;
+        let response = self.client.get(&url)?;
         let data: CapitalChangesResponse = crate::rest::read_json(response)?;
 
         Ok(data)