+++ core/src/rest/stock/snapshot/quotes.rs
@@ -71,10 +71,7 @@ impl<'a> SnapshotQuotesRequestBuilder<'a> {
         }
 
         // Make request
-        let request = self.client.agent().get(&url);
-        let request = self.client.auth().apply_to_request(request);
-
-        let response = self.client.execute(request)?;
+        let response = self.client.get(&url)?;
         let quotes: SnapshotQuotesResponse = crate::rest::read_json(response)?;
 
         Ok(quotes)