+++ core/src/rest/stock/technical/kdj.rs
@@ -127,10 +127,7 @@ impl<'a> KdjRequestBuilder<'a> {
             url.push_str(&query_params.join("&"));
         }
 
-        let request = self.client.agent().get(&url);
-        let request = self.client.auth().apply_to_request(request);
-
-        let response = self.client.execute(request)?;
+        let response = self.client.get(&url)?;
         let kdj_response: KdjResponse = crate::rest::read_json(response)?;
 
         Ok(kdj_response)