+++ core/src/rest/stock/technical/rsi.rs
@@ -93,10 +93,7 @@ impl<'a> RsiRequestBuilder<'a> {
         }
 
         // Make request
-        let request = self.client.agent().get(&url);
-        let request = self.client.auth().apply_to_request(request);
-
-        let response = self.client.execute(request)?;
+        let response = self.client.get(&url)?;
         let rsi_response: RsiResponse = crate::rest::read_json(response)?;
 
         Ok(rsi_response)