+++ core/src/rest/stock/technical/sma.rs
@@ -93,10 +93,7 @@ impl<'a> SmaRequestBuilder<'a> {
         }
 
         // Make request
-        let request = self.client.agent().get(&url);
-        let request = self.client.auth().apply_to_request(request);
-
-        let response = self.client.execute(request)?;
+        let response = self.client.get(&url)?;
         let sma_response: SmaResponse = crate::rest::read_json(response)?;
 
         Ok(sma_response)